# docs/swagger.yaml
basePath: /api
definitions:
...
  response.Response:
    properties:
      data: {}
      error: {}
      message:
        type: string
      success:
        type: boolean
    type: object
  sections.CreateSectionRequest:
    properties:
      display_order:
        minimum: 1
        type: integer
      duration_minutes:
        minimum: 1
        type: integer
      section_code:
        maxLength: 50
        minLength: 3
        type: string
      test_id:
        type: string
      title:
        maxLength: 255
        minLength: 3
        type: string
    required:
    - display_order
    - duration_minutes
    - section_code
    - test_id
    - title
    type: object
  sections.UpdateSectionRequest:
    properties:
      display_order:
        minimum: 1
        type: integer
      duration_minutes:
        minimum: 1
        type: integer
      is_active:
        type: boolean
      title:
        maxLength: 255
        minLength: 3
        type: string
    required:
    - display_order
    - duration_minutes
    - title
    type: object
...
host: localhost:8081
info:
  contact: {}
  description: Backend API for English proficiency testing platform.
  title: Go English Test Platform API
  version: "1.0"
paths:
...
  /sections:
    get:
      description: Get list of sections
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: Get all sections
      tags:
      - Sections
    post:
      consumes:
      - application/json
      description: Create a new section
      parameters:
      - description: Create section request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/sections.CreateSectionRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Create section
      tags:
      - Sections
  /sections/{id}:
    delete:
      description: Delete section by ID
      parameters:
      - description: Section ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Delete section
      tags:
      - Sections
    get:
      description: Get section detail by ID
      parameters:
      - description: Section ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
      summary: Get section by ID
      tags:
      - Sections
    put:
      consumes:
      - application/json
      description: Update existing section
      parameters:
      - description: Section ID
        in: path
        name: id
        required: true
        type: string
      - description: Update section request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/sections.UpdateSectionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Update section
      tags:
      - Sections
  /sections/by-test:
    get:
      description: Get sections by test ID
      parameters:
      - description: Test ID
        in: query
        name: test_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      summary: Get sections by test ID
      tags:
      - Sections
...
schemes:
- http
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"